[INTERNAL] Demo Kit: Version select dialog now consists of a tree
After a user feedback we decided to change the version select dialog
in order to improve readability and UX.

Previously the dialog content was a list which was hard to read.

Now the dialog content is a tree. This way we have way less items/branches
initially. If after filtering (user input in the search field) there
is only one branch left - we expand it.

JIRA: BGSOFUIPIRIN-5121
Change-Id: I3c0abb2b83f9d3ad3dbed1bb5803cd6ffdf2de9b