[FIX] sap.uxap.HierarchicalSelect: Fixed button's focus on XL desktop sreen
With I5db4521d291bea51ff67d80e0d67b8043d75286d a new breakpoint DESKTOP_XL in ObjectPageLayout
is introduced and the CSS query for applying to proper height to the HierarchicalSelect
IconOnly button is not wokring. With this change we add the new breakpoint class to the query.

Change-Id: Ifb1917640bb6a49adf1992cac84d121037d85853
BCP: 2180251671