[FIX] JSDoc: fix pseudo-fields, declare type for real fields
When class methods are copied from a mixin or other method name,
often the @function tag was missing, making the method mistakenly show
up as a field in the API reference (fixed by adding @function).

'Real' fields should have a @type and should be marked as @readyonly
when they act as a constant. Note: @const can't be used as it implies a
new kind in JSDoc, something which is not yet handled by our SDK.

Fixes: https://github.com/SAP/openui5/issues/3294
Change-Id: Ic058dbf0f30469c3de479404b44d8bcabe85c4dc