[INTERNAL] Field: On cloning use Link settings of template
If a Field with FieldInfo is cloned the clone should initially have the
same Link settings like the template. That means, if the template
already knows that a sap.m.Link should be rendered this is taken as
default on the clone. Only if the FieldInfo later is not triggerable a
sap.m.Text is created. 
This is made to prevent the creation of unneeded sap.m.Text controls
inside a table if sap.m.Links should be shown. As default the setting of
the template should fit in the most cases. 

Change-Id: I57c41c0bc3566bae9f2cd4564c195770ad61177e