[INTERNAL]ListView: enable count for first column
The SmartTable prior 'P13nColumnsPanel' used to show the count of
selected columns in comparison to the available columns.

As the 'ListView' was now being used, this functionality was not
available anymore. With this change the count has been added as
feature on the 'ListView' to enable the count similar as before.

E.g: "Columns (3/12)" instead of "Column"

Note: the flag needs also to be set to true in comp in a separate change.
This feature could now also be added to the 'Adapt Filters' Dialog.

Change-Id: I8e655f36d17aa920a1b0a0337052fbd9a5de2796