[FIX] webc.fiori.FCL: fix default value for property accessibilityTexts
In UI5 Webcomponents, the default value for properties of type 'Object'
is an empty object. The implementation of FlexibleColumnLayout relies on
that default value and assumes that the property accessibilityTexts is
non-nullable.

In UI5 Classic, the default for properties of type "object" is null. If
no defaultValue is specified in the metadata, the defaultValue becomes
"undefined". Since change I5a8bfaf5ad3433b332bba3782664e5aa19c098f6,
properties of type "object" are copied 1:1 during onAfterRendering,
which now causes the FlexibleColumnLayout to crash as it can't handle
the "undefined" accessibilityTexts.

Setting the UI5 Classic defaultValue to {} avoids the immediate problem
but still errors might occur when a caller sets null or undefined as
property value.

Change-Id: I9ec4e6443cc223cd87ea37eded72f697286542e6