[FIX][INTERNAL] sap.m.ProgressIndicator: Fixed text contrast ratio
Problem:
When the text is in the left side of progress indicator it doesn't have
enough contrast with the background.

Fix:
The issue was synced with visual designers and the value of parameter
require more resources to be changed since it is used in many controls.
Their suggestion was to replace the threshold only for sap.m.ProgressIndicator
at the moment with following values:
- belize - 0.6
- fiori_3 - 0.5

The default value of the parameter is kept for all other themes.

BCP: 2180165340
Change-Id: I53be1d0cd0622280d0f84ab0cdd55142f5b0f1d3