[INTERNAL] Table: remove wrapping style for label
The sap.m.Label control already supports wrapping, so its no longer
necessary for the table control to influence the wrapping for the labels
in the table.
If wrapping is desired the consumers should enable wrapping for the Label
control and the Label control can take over the wrapping behavior.
BCP: 2180079113

Change-Id: Iaa69d53e97baacd539bc73e418d344f1193a2f69