[FIX] ResourceBundle: Improve assertion message when text can't be found
The assertion message only showed the bundle url (without locale), which
can be confusing in case the default bundle is not part of the
supportedLocales.
A user might think that the file has been loaded, and in case the key is
actually part of the bundle, it looks like a wrong behavior of the
ResourceBundle.

BCP: 2180263498
Change-Id: I77e7fdc9f6792ab08738fa6e141b233d518f4060