[FIX] sap.uxap.ObjectPageLayout: Fix infinite loop error
Problem: When visibility of one of the controls in "blocks" aggregation
of the SubSection is changed, onBlocksChange event is fired. It
calculates again the layout, in order to visualize correctly the blocks.
However, in _applyLayout function the blocks are added again in the
inner Grid as "content" aggregation and this triggers modelContextChange
event is some cases. Problem here is: the application is subscribed for
the modelContextChange event and every time it switches the visible
property of the block from "true" to "false" and then again to "true" or
"false". This triggers new onBlocksChange event for the
ObjectPageSubSection, new recalculation of the layout, new
modelContextChange event for the application and results in an infinite
loop.

Solution: in _applyLayout function now there is a check if the block has
already been added to the "content" aggregation of the inner Grid and is
not added, if it is already there. So no modelContextChange event is
fired for the application.

BCP: 2180239658
Change-Id: Iac2ac28d9ca4ca03e1ce6fc2af987cac02a2f81d