[INTERNAL] sap.m.SelectList: Removed border overwrite
By the latest visual design specification, the sap.m.Select's
dropdown (sap.m.SelectList) should have equal border radius on all of
its corners like it is in the sap.m.ComboBox's dropdown.

This is why we remove the redundant border radius overwrites in all
of the themes.

JIRA: BGSOFUIPIRIN-5322
Change-Id: I9c922d2147c5ab54730af097e1bd7c73f9219f72