[INTERNAL] sap.f.DynamicPageTItle: Correct _getARIALabelReferences function
The title of the heading is not always a control but only a string and the function call to get the id was failing when concatenating the values for aria-labelledby attribute. Check if getId function exist on the oTitle variable before calling it.
Caused by: Iaacaa9effd8b8ccd773d43ce0c70ee044c2c4748

Change-Id: I58997873b2da18d9357646c50f22f5f85fdfd5b2
BCP: 2180298082