[INTERNAL][FIX] mvc.Controller: Fix undefined _aDestroyables in case of missing super call
If an applications Controller implementation is missing the mandatory
constructor super call, the list of managed 'destroyables' was undefined.

With this change we log an error, but for compatibility reasons keep
the existing behavior stable.
Due to many application failures a clean super call chain cannot be
expected or enforced.

BCP: 2180298619
BCP: 2180301252
Change-Id: I1408bfe174eb88d60c0b0558f882cf27a6dc2818