[FIX] sap.ui.VersionInfo: copy 'hasOwnPreload' flag for components
When the sap-ui-version.json contains the information that a component
is packaged separately from its containing library (property
'hasOwnPreload'), then the sap.ui.VersionInfo should propagate this
information when asked for the dependency information of that component.

Before this change, the property 'hasOwnPreload' was omitted from the
returned dependency info.

Change-Id: Ia5cd3b7fe4f61a21949bd709ea1a07c5c706115a
BCP: 2180290030