[INTERNAL] sap.m.Avatar: Improved ACC sample page
Problem: Active avatar elements don't have label.

Solution: We add 'ariaLabelledBy' with a test label as property to
all of the active avatars on the page.

Additionally the sap.f.Avatar (deprecated) ACC sample is now replaced
with a sap.m.Avatar one. This way we support only one file.

Also the visual test is transferred from  the "f" to the "m" library.

BCP: 2180274631
Change-Id: I2ba464e43023eaefb02f4f566d5dfd18db810d6f