[INTERNAL] ConditionType: parse number part of unit to null
if no number was already used. undefined is not a valid value for parts
of the Unit or Currency type.
So if the Field is initally empty and the user enters a unit (or a
number) the other part is initialized with null. Otherwise the
formatting of the value fails.

Change-Id: Ib78313de42b8cddc8ba9d3d36be6c55c0f8e591d