[INTERNAL][FIX] sap.ui.rta: Add Elements evaluating wrong property name
In cases where an OData property has a name that starts with the name
of another property (e.g. "Product" and "ProductBaseUnit"), it was
possible that the incorrect property was evaluated to be available for
revealing.
For example: "Product" could have annotation UI.hidden = true
and another property "ProductBaseUnit" would also not be available on
the dialog just because its name starts with the other property's name.

Solution:
Improve the existence check of the property by considering "/" in the
name of navigation properties.

Change-Id: Iee605d4541db933eead7a5b9f04fb86854985cb8
BCP: 2180306129