[INTERNAL][FIX] sap.ui.rta: Add Elements evaluating wrong property name
In adaptation mode, adding a field, removing it and trying to add it
again fails because the wrong property is being evaluated.

Additionally:
In cases where an OData property has a name that starts with the name
of another property (e.g. "Product" and "ProductBaseUnit"), it was
possible that the incorrect property was evaluated to be available for
revealing.
For example: "Product" could have annotation UI.hidden = true
and another property "ProductBaseUnit" would also not be available on
the dialog just because its name starts with the other property's name.

Solution:
Improve the existence check of the property by considering "/" in the
name of navigation properties.

Change-Id: I921a9664f8c2d31a59dd65e7803915774aa5aa93
BCP: 2180306129
CR-Id: 002075125900003905912021