[INTERNAL] sap.ui.dt: move performance folder to internal/performance
as in fl and rta the performance folder should not be part of the bundle.
the test apps use inline styles and inline scripts, those should not be delivered with ui5.
Also deleted some old test apps that are not in use anymore

also changed a test that used inline styles

JIRA: TEAM42-135
Change-Id: I843820522303f8fabd21756fea03c97a9bb5261f