[FIX] sap.f.ShellBar: Stabilized qunit
Background:
The test cheks for initial requirements: that its items are initially
rendered for desktop size (as qunit-fiture width is 1000px which is desktop size).

However, the initial setup takes longer to complete because the 1000px
of qunit-fiture width is not enough to render all items and 2 items need to overflow
=> the setup is complete once the inner OverflowToolbar
completed the layout of its content.

Problem:
There is no suitable event from the Shellbar to notify
that the inner OverflowToolbar completed the layout.

However, the layout just adapts to the default qunit-fiture width (1000px),
which is smaller than needed to render all initial shellbar items.

Solution:
As a simple solution, emsure that by default the shellbar
(and the inner OverflowToolbar) have enough space to render its items
=> the consequent re-layout (to add two items to the overflow)
is no longer needed
=> the initial conditions setup completes earlier

Change-Id: I30569789b1e57f236868b5a5eaad9535d05c68ac
JIRA: BGSOFUIPIRIN-5308