[FIX] sap.m.Input: List arrow key navigation behaviour fixed
Suggestion List items should update input's value differently based on
ListItem's type.
- If there's a "startsWith" matching item, complete the item and highlight
 only completed part. When the focus goes to the input, reset to user's input
  value.
- If it's a "contains" item, complete the item, but hightlight the whole text
in the input. When the focus goes to the input, reset input's value to an
empty string
- If it is a GroupHeader item, set input's value to an empty string. When the
focus goes to the inut, reset to input's value to an empty string.

Note: If the first suggestion item is "startsWith", but the next ones are
GroupHeader or a "contains" item, the value of the input would be reset to
an empty string.

BCP: 2180308821

Change-Id: I04856eccfef468b0b54aeeb3d372cc2469fff88e