[INTERNAL] sap.f.DynamicPage: Update DynamicPageTitle padding earlier
Problem: Padding is applied to the DynamicPageTitle later than the
scrollbar appearance, which causes additional delay of performance
scenarios.

Solution: The padding was updated with a delayed function, after the
scrollbar appearance. Now the _updateScrollBarOffset function is deleted
and all styles are just added in the .less files. There are some
differences in the native scroll width in different browsers (FireFox is
different than Chrome and Edge) and OS (FF on iOS is different than FF
on Windows). That's why in .less styles we take into account the
differences. What is more, padding is only added on Desktop (no
scrollbar on Tablet and Mobile device) - so this logic is also reflected
in the new styles. Tested with RTL - also works fine now.

JIRA: BGSOFUIPIRIN-5336
Change-Id: I6021bd6514b2fcbbc75fd37e7e3905019e398c7e