[INTERNAL] DefineConditionPanel: on switching operator - empty token
If a condition with value is changed from equal to contains an empty
token is created. The value is cleared. 
As for some types (like String with isDigitSequence) a special parsing
is needed the value needs to be formatted and parsed again. But as it is
not known what the user initial entered it could lead to a different
value. 

Change-Id: I68efe220f7ee8c5028f9a95cab0d408df2906ade