[INTERNAL] Demo Kit: Bracket app samples inside iframes
With the development of Demo kit Bracket App we want all the resources to be loaded,
independent from demo kit application. This is impossible when we want to load samples
inside demokit app and bundle them with different version of ui5, unless we load all the
samples inside an iframe.

Change-Id: I8c29ee940555c6b0bad2d5a8969e58547378d7bd