[FIX] sap.uxap.ObjectPageLayout: Corrected subSection height
Problem:
When the header is expanded in the content artea, the height of subSections
with "sapUxAPObjectPageSubSectionFitContainer" class is maller than expected.
The difference from the expected height is the height of the anchorBar.

Solution:
The logic that calculates the required height [of the subSection]
now uses metrics that do oot depend on explicit check for presense of anchorBar
(it now substracts subSection.offsetTop - container.offSetTop to obtain the
height of the content above the subSection, regardless the content itself)

Change-Id: Iabdc18c619800f015c776a07ade67db0dd976e5d
BCP: 2180174424