[FIX] sap.uxap.ObjectPageLayout: Correct subSection height when header in title area
Problem:
For case where the header is expanded in the title area, the height of subSections
with "sapUxAPObjectPageSubSectionFitContainer" is smaller than expected.
The difference from the expected height is the height of the anchorBar.

Solution:
Ensured that the enchorBar height is excluded from the computation
of the height of the sections area.

BCP: 2180174424
Change-Id: Ibdbd1ccdb086c9ce11aaa1e4c0c2f17716dbb253