[FIX] sap.m.Popover: Resize properly Popover after open
The Popover needs to adjust its size after opening, so that the content
would fit inside.
With the fix for the race condition (Ia13bd78fc9422b18cfd2261d9483cf4453ca8e53)
the resize handler triggers only for Popover's opening, but not for OPEN.
We'd need to delegate size adjustments to happen on after opening of the
Popover.

BCP: 2180252966

Change-Id: I4fd7f48cbdaad14fccc76bd6ff054e4db6cc4c65