[FIX] sap.m.App: Fix error when App is invisible and has no parent
Problem: Error is thrown, when sap.m.App is rendered initially with
property visible="false" and has no parent. The problem comes from the
logic, which tries to set 100% height and of the parent DOM elements
of the App.

Solution: Check if the sap.m.App has its own DOM reference, before
trying to traverse the parent elements.

BCP: 2180221467
Change-Id: I785754bb20db6b3f25c7cec2888e8a1f659d88c4
CR-Id: 002075125900004105602021