[FIX] sap.m.Tokenizer: Calculations based on density mode
The Tokenizer now takes into consideration the density mode of the UI
in order to calculate the width of the n-more popover.
This solves the issue where long text of some tokens will get
truncated when shown as list items.
The difference comes from the fact that the Token and the ListItem use
different font size, thus the ratio needs to be taken into account
when setting the Popover's content width.

BCP: 2180255274
Change-Id: Ibfd0335c192c7f9f3129800f0793e5d7cf361621