[INTERNAL] sap.ui.core.hyphenation.Hyphenation: Use browser version 3.4.0
Hyphenopoly now provides browser version with "Hyphenators" feature.
The hyphenators allow to programatically hyphenate text for a given language.

The purpose of this change is to simplify the usage of the third-party library
(unmodified and unpatched).
Also to enable CSP compliancy - instead of wasm to use asm.js
(This is the last Hyphenopoly version that has asm.js support)

Change-Id: I2255454a31a81def94c88c3b651c83a8f0dd0c67
JIRA: BGSOFUIRODOPI-1860