[INTERNAL][FIX] sap.m: Support MenuButton "Split" mode in CombineButtons handler
When combining two buttons via RTA a MenuButton is created by the CombineButtons
change handler. If this MenuButton is set to Split mode, actions are not
necessarily triggered from the menu items.
Therefore the change handler needs to attach to the menu events instead to
still fire the original button events after an action was pre-selected.

Change-Id: I9ee2312d371b1e8f7b20a57fe38dcefa89236224
BCP: 2180298764