[INTERNAL] FrameOptions: Resolve TODO for eslint rule
This has been detected by enabling the no-unmodified-loop-condition
rule.

Currently, not all scenarios for the affected lines are covered by tests and
the behavior cannot be easily tested with unit tests. Additionally,
manual tests would be required. Because the current code is working, and we
might risk a regression, we decided to not simplify the loop without
real need.

Change-Id: I797eba8c33c309187259a961544b9298f995e48f
JIRA: CPOUI5FOUNDATION-379