[INTERNAL] sap.m.PlanningCalendar: has consistent focusing on navigation or date selection
Issue: when a date is selected from the picker or there is a navigation
performed, the PlanningCalendar was changing the date property to the
currect row instance, which was causing unwanted focusing of the corresponding
date element in the UI.

Solution: now the PlanningCalendar is using the displayDate functionality
of the row instance to display a date without setting the focus to it.
As result the focus after the interaction stays on the UI element,
which has triggered it.

Change-Id: I6d7f987f8dc8dfcb0e54f3bbe9fa5b895b003f6c