[INTERNAL] sap.ui.webc.common.WebComponent: improvements
This change addresses several deficiencies in the WebComponent.js base class,
found during testing and the integration of new components.

1. Event parameters conversion - limited to 2 levels of depth.

This prevents bugs that occur when big objects or objects with recursive references to each other
are passed as parameters to events (which causes endless loops).
Realistically event data that contains DOM references, that need to be interpreted as
OpenUI5 controls, consists of an array or an object or a 1-level-deep combination of these.

2. Support for rendering attributes/properties of custom types added.

Until now the WebComponent.js/WebComponentRenderer.js code assumed that
only properties of "type: Object" could be objects. This is however not the
case. Custom property types can be anything. An example is a custom type that
accepts either a string or an array of strings (so string or object).
Therefore the logic that decides whether a property must be set declaratively
(via attribute during render) or imperatively (during onAfterRendering), now
checks not only if the type is "object", but also whether the value itself is
an object.

Change-Id: Ie23b5cca91b8ffdb34dda1d58a046fa54d94b380