[FIX] sap.m.TimePicker: fixed flickering of the popover scrollbars when height is reduced
* like change #5371792 (where the same problem is fixed for DatePicker),
there was flickering of the scrollbars and popover size when height of
the screen is reduced and the clock inside the popover cannot be displayed
without scrollbar. This was caused because in described case first a
vertical scrollbar appeared, which reduces width for the content (calendar),
and this was followed by showing of horizontal scrollbar.

* now this is fixed and there is no flickering.

Change-Id: Iafd13c2b3ed8ba01e59f1ed958fd9eae7ea396a7
BCP: n/a