[INTERNAL]P13nBuilder: focus handling after reset
After pressing reset the 'reset' Button will remain disabled (in V2)
In MDC it will remain on the 'reset' Button as in V4 the implementation
is different in terms of the state handling of the perso controller
implementations.

To streamline this behaviour and to avoid focus lost, the focus
will be set to the "OK" button after a reset has been confirmed
(if its canceled it will remain on the "reset").

The "OK" button will now be focused as most likely action after resetting
the personalization, similar to the earlier P13nDialog implementation.

BCP: 2180324357
Change-Id: I5dd4b220b2e6b5bc5a7bdc5f6ebe9e6a2656ad86