[INTERNAL] sap.ui.rta: Show all possible entries when revealing elements
Elements hidden in one aggregation are not displayed as available
for reveal on a sibling aggregation, even if moving the revealed element
to that aggregation is possible.

With this change, all elements that can be revealed and moved to the
selected aggregation will be visible on the available elements dialog.

Change-Id: I7d353098224b1e5ba90a1c44dc97e0aee440312d
JIRA: TEAM42-122