[FIX] sap.f.DynamicPage: Correct maxScrollHeight computation
Problem:
The page computes that it needs a scrollbar when it doesn't
because of 1px maxScrollHeight computed as a result of browser rounding.

This is a regression from a previous change I4ced0e1ed1c724600a2ef985b8e60d2f4b84a106
that adresses the same issue by switching to use precise height
(obtained via getBoundingClientRect) instead of using the
integer clientHeight, because the browser returned a floored value
for clientHeight (smaller than the precize height).

However, in the new use case, the browser returned clientHeight
that is 1.1px bigger than the precize height
=> redundant scrollbar because of 1px scrollable content computed
after comparing scrollHeight with the precize height

Solution:
Ensure that when comparing scrollHeight against clientHeight,
we compare against the maximum obtained clientHeight
(to avoid ending up with 1px difference
in favour of the scrollHeight because of rounding).

BCP: 2180305489
Change-Id: I0850508f33649b93981e279c6e554b7de17912c1
(cherry picked from commit 50de284246ff115384e4c26eceb30ca92641c87c)
CR-Id: 002075125900004771662021