[INTERNAL]Personalization: avoid tooltips
The P13nBuilder is used to create the outer p13n Dialog (not the
inner panels) and the underlying JSON Model structure.

The P13nBuilder implemented the V2 fallback logic to provide the "label"
as fallback in case no explicit tooltip has been provided.
The fallback has been removed and the tooltip will only be shown in case
it has been explicitly provided.

This change affects:
- MDC Table "View Settings" Dialog (+ SmartTable)
- MDC Chart "View Settings" Dialog (+ SmartChart)
- MDC FilterBar Adapt Filters Dialog (+ comp FilterBar "Adapt Filters")

Jira: CPOUIFDENVER-1672
Change-Id: Ia22e693e40f272bd97178abaa2d1f00d3198e242