[INTERNAL][FIX] sap.ui.fl: Properly encode IFrame src URLs
- Encode IFrame src URLs before rendering to include binding values
- Avoid encoding parts of the URL before saving the change to prevent
duplicate encoding
- Resolve binding values before validation
- Validate the URL when adding parameters from the predefined table

IFrame source URLs support binding strings which can contain characters
that need to be encoded, e.g. whitespaces.
The recently introduced URL validation will reject such values if they
are not properly encoded, which is a critical regression that can break
existing customer changes.
Before this fix, the AddIFrameDialog was responsible for encoding URLs
except the parameter values. This means that the non-parametric section
would currently be encoded twice, which needs to be changed as soon as
possible to avoid irreversibly broken changes.

Change-Id: I63b8359ecceea6ecade905fa8467c49f87bb0898
BCP: 2180345725