[FIX] sap.m.Dialog: Dragging when the page is scrolled now works
On initial click on the header of a draggable dialog, the x and y
position are now taken relatively to the viewport, which works even if
the page is scrolled.

This situation occurs if a parent itself is too tall, causing the page to have a scroll.
Does not happen with most container controls, which handle their scrolling with an explicitly
set height (for example height: 100%) and overflow: auto.

Change-Id: Ic20fd45da8597680d5c80218aa66893b82e2f5b9
BCP: 2180321625
Snippix: 57038