[INTERNAL] sap.m.Select: Correct tooltip support
Problem: Tooltips was not shown for the ordinary select control, but only in the
icon only mode.

Solution: Add toolitp to the label element as well.

Note: The title attribute on the focused element is not showing because of the sapUiPseudoInvisibleText class.

Change-Id: I15893396a18e08fd5b5cc97de7890128ca212b9a
BCP: 2180324353
BCP: 2170194256
(cherry picked from commit 627eae0f975a7f29036e2e27a5695ddf8603c80d)
CR-Id: 002075125800002470922021