[INTERNAL][FIX]sap.ui.rta: Prevent Context Menu opening after rename
Pressing "enter" after a rename on an Overlay was causing the Context
Menu to come up. This is because the Overlay got focus after the rename,
and pressing "enter" on an Overlay opened the MiniMenu.

This change flags the Overlay when a rename happens, so the next
ENTER keyup event is ignored and the context menu does not come up.

Change-Id: Ib9fdf8946e229a4086ac9566fe26c621b14d24b2
BCP: 2180254975