[INTERNAL][FIX]ConditionFlex: wait for change appliance
With the recent CSP changes in sap.ui.fl most modifier methods now return
a promise, such as:
- Modifier#getProperty

When applying multiple changes at once, e.g: [oChange1, oChange2, oChange3]
The changes will be processed sequentially. In case of FilterFields
in the FilterBar use in/out parameters, there are multiple changes
provided individually in via condition model, each generating a singular change
and not a set of changes which causes the changes to be processed in parallel.

Before the async replacement this was not an issue, as getProperty synchronously
retrieved the current filterConditions - as this now is asynchronous the
changehandler needs to ensure that he waits until he retrieves the filterConditions
as they might have changes inbetween.

Change-Id: Iff421bb138b142d0a8b820309da29df4810d9643