[INTERNAL][FIX] SelectionModelSelection: Avoid unnecessary selection deletion
If something is selected between initialization and the first binding change
event, the selection should be kept. This fixes a bug where the
SelectionModelSelection plugin would clear the selection when the first change
event is fired.

The reason for this issue is that the table detects a binding length change
(from 0 to >0) after the binding fires the first change event. The plugin is
responsible to clear the selection when the length changes, because of the
deficits of the index-based selection. The initial change is now ignored.

Change-Id: Iba885411cb999e063eb27de5e5c9c45e8344e651
BCP: 2180349712
CR-Id: 002075125900004511522021