[INTERNAL][REVERT]GroupView: ensure filtering does not destroy fields
This reverts commit efbbdbf4978b5fbd5bf456aca0d40781af2ef491.

Reason for revert: The change was meant to prevent accidental filter field
destroying, the change introduced a different side effect as the update is
slightly delayed due to the model binding update, which might cause that
the filter field is updated to the GroupView as the same instance is being
used between both views, causing the ListView to be displayed without filter
fields.

A different solution needs to be implemented for destroy related issues,
ideally the binding should not destroy the FilterFields.

BCP: 2180324633
Change-Id: Id3a21c4cf9f3e8034b57058a1eeb1143746d456d