[INTERNAL] sap.f.semantic.SemanticPage: One visible action shown
When there is only one visible action (default or custom), the Share
menu icon should not be shown. Instead, the action itself should be
rendered in the title toolbar and it should trigger direct action.

Default actions should be shown only as icons and the custom ones are
shown as they are specified by the user.

In addition, when no visible actions are available, the ShareMenu icon
is not shown now. Previously it was shown, but when opened, the Menu was
empty, which was not OK.

JIRA: BGSOFUIPIRIN-5351
Change-Id: I18bf282e36013c817c75b6b33f2c7d41243eea90