[INTERNAL][FIX] sap.ui.fl: Safeguard undefined revert data in custom data
When trying to save the revert data of a change as custom data to use
it with the view cache, it can happen that change handlers return undefined.
In such cases we need to make sure that we don't save the revert data
as it will lead to app crashes when later trying to parse it.

Change-Id: Iaf9709d8662cc59bbacf8be8a6dc8fddc7f111e6
BCP: 2180356692
CR-Id: 002075125900004635312021
(cherry picked from commit 6c224ecd6905a8836809d9b42becfbb7b36f64ac)