[INTERNAL] FESR: fix wrong step names
The first control event after an interaction relevant browser event
starts an Interaction. We determine the step name for FESR out of
the trigger (controlID) and event (name of control event).
In some cases this event does not match the action the users triggered:

- User edited some field (focus in this field)
- User clicks save

If the the focus movbed to the button due to the click the field
fires some events to e.g. validate the value. Afterwards the click
on the button is reported.

The reported step name in FESR is not 'save_press' as expected, but
something like 'validation_success'.

This issue is adressed in this change.

Change-Id: I8c02b61a8d55f4c35d13749ea636c6dc3a23da20