[INTERNAL][FIX] Table: Keyboard navigation in nested ItemNavigation
If the table is nested in a control that uses the ItemNavigation, both controls
react to the same event, because the table does not stop propagation in
all cases. To avoid navigation issues, the table mimics the behavior of the
ItemNavigation and always prevents default and stops propagation on events that
might change the focus (not necessarily do change the focus).

Change-Id: I12fba760593760495c7160d04970623c068f7697
BCP: 2180308670