[INTERNAL][FIX]ItemBaseFlex: cleanup promise chain
In case the application is changing its baseline (for example a column
has been added after changes have already been created for this column)
The changehandler should react gracefully and not break the promise chain.

E.g. the change should be marked as not applicable and the change processing
should continue.

In case an item is already present in the control while it is tried to be added,
the changehandler did not correctly handle this case which has been resovled in
this change.

BCP: 2180358779
Change-Id: Ia5247e6d9a0ed461f908cf9de4a1fea27258e436