[FIX] sap.m.InputBase: Display initially bound value when in read-only mode
Previously when the Input was read-only or disabled the initially set
selectedKey in the view, didn't update the Input's DOM as it was not yet
rendered. It was only possible to set it after the initial loading via
the setter. Now the value is displayed properly after the loading
of the control when there is an initially set selectedKey.

BCP: 002075129400005050852021
Change-Id: Ibe2366d84386f3ff11e4de8edd9504a6ff3940cd