[INTERNAL] sap.m: CSP fixes for controls in m library
This change removes hardcoded inline styles for certain controls in order
to comply with the most strict CSP policy.
The issue is not that there are inline styles, but that there's an initial
string which would later be parsed into DOM and it contains inline styles.
If styles are applied later over a DOM element(s), there won't be issues with CSP.

Covered controls:
- RatingIndicator.js
- MessageView.js
- ScrollContainer.html

JIRA: BGSOFUIRILA-3181
Change-Id: Ide8dcd6c576e4c6e35c91d5e6f5ae2c8a57ac4fc