[INTERNAL] ListBase: fix range selection
1. When there are some hidden items (visible=false) and rangeSelection is
performed via selection 1 item and then SHIFT + mouse click, the rangeSelection
object created by the ListBase was incorrect since when creating the object
visible items were taken into consideration and when selecting the range
the indexOfItem() was used. This caused inconsistency for selection.
The RangeSelection object is now created using indexOfItem to be consistent
and solve the issue.
2. When includeItemInSelection=true on List, this allows selection of the list item
by tapping on the item. The rangeSelection did not detect this mode so far
and rangeSelection was only possible by tapping on the checkbox. This issue
has now been fixed
BCP: 2180376172
BCP: 2180311141

Change-Id: Ie090da85d815ee8eea7390354c3daac793367ea6