[FIX] sap.m.MultiComboBox: 'selectionChange' is fired properly on 'selectAll'
This commit - caad29f3b96fe7ae437875c1fda102b857e19361 introduces a
backward incompatible change related to the select all (CTRL + A
keyboard handling) functionality of the control.

Previously when all items are selected via CTRL+A shortcut the
'selectionChange' event is fired as many times as there are items
selected/deselected. The change introduces new parameter to the event
which contains all items in an array and the event is fired only once,
but this will break the applications that already rely on that event
for that interaction.

Now when the selectAll keyboard combination is used the event is fired
for each selected/deselected item.

This commit fixes the incorrect 'selectAll' behaviour that selects all
items no matter if they are filtered or not, now only the filtered items
get included in the selection.

This change also fixes the bug when 'selectAll' is used on a GroupHeader
or SeparatorItem - the group headers are no longer tokenized, only the
list items.

BCP: 2180358601
CR-Id: 002075125800001964082021
Change-Id: I724dfd1bb39a82ffcc769f923f136c8fc74ac214