[FIX] Demo kit: Fix neo-app.json path and sample iframe path
Neo-app.json file exists in previous versions, but it doesn`t contain
'routes', which is required for the change version button to appear.
This is the reason, we also need to load it from root URL along with
versionoverview.json. Using the proxy, we handle this logic internally.

In index file used for samples to load in iframe we need to exclude
special characters in order to not be passed inside a query param to the
iframe.

BCP: n/a
Change-Id: I918f543ed43fa37f48454ce73d5c3188afb27b74