[INTERNAL][FIX] Table: Positioning of the row actions
When the column widths contain subpixels, the full width of the table and the sum of the column
widths might differ for a fraction of a pixel. This affects the positioning of the row actions and
the same are not in the visible area. Therefore the calculation is adjusted with +1 pixel.

Change-Id: Ia614288a6680ce43092daf2230cfec587ea5cce2
BCP: 2170172514