[FIX] sap.m.StepInput: formats its value correctly
When the displayValuePrecision is changed to a wider one, but it is still
narrower than the passed value, the value was getting cut.

Example: When the value and dislayValuePrecision properties are bound to a model,
the StepInput is initiated with it's default value - 0 and the value precision
is calculated to be 0. Than the value is set to 1.23 and the displayValuePrecision
is set to 1 from the model. The result was "1.0" instead of "1.2".

Solution: The original value - the one before formatting against
displayValuePrecision, is saved in an internal variable and then used for further
formatting.

BCP: 002075129500006083982021
Change-Id: I4dd6ecd693ad433040734ecf884763332c108aaf