[FIX] sap.m.OverflowToolbar: Racing condition with invalidating content fixed
When we change a property of a control inside the associative popover and that property
triggers invalidate the closing of the popup, triggered by _resetToolbar happens before the
invalidation and the closing called by AssociativePopover itself, which leads to
losing of the focus of the OTB button, focused before opening of the popover.
There is a logic deleted with a change b731561abb2c40e3d92913c2074bd82eb7f5b4e5 in which
we force the focus back to the OTB button after closing the popup.

BCP: 2170174725

Change-Id: Ib14e4d00a7df07946ee1c5d4755701c9f3d396fa