[FIX] Table: Column header double announcement
JAWS announces the header twice, because the gridcell points to the columnheader
twice, with the "headers" and the "aria-labelledby" attributes. The "headers"
attribute is removed, because this seems to have less negative effects than
removing "aria-labelledby".
There are still issues in multi-header scenarios and when navigating between
fixed and non-fixed columns. JAWS is not very cooperative in these cases.

Change-Id: I4e6fb6f6be13f0946521a59f8db413753485f7e3
BCP: 002075129400005585052021