[INTERNAL] mdc.util.PropertyHelper: Adaptation of allowedForComplexProperty and ignoring of attributes
Having a PropertyHelper with a deep and detailed structure that should be supported also for complex properties each single property must be marked with 'allowedForComplexProperty: true'.
To don't do this it would be sufficient to just define the parent object with 'allowedForComplexProperty: true'.
It is still possible to exclude a single property within a child node with 'allowedForComplexProperty: false'.

It is also now possible to ignore a PropertyHelper attribute from defining its default completely if desired.
In that case the entire attribute will be defined as null.

The idea of this is change comes up with Change-Id: I82744fcc1f8ed6e80881b451e0fd117e85b8b9e6.

Change-Id: Iac29e551480a7fe9032bc98dbd6ad4727b284f28