[INTERNAL][FIX] sap.ui.fl Correct restore external data in CompVariantMap after invalidate cache
If the VariantManagement is completely blank (with only Standard
variant), after invalidate cache, the entry of it in the map is
completely deleted without re-initiate. So the entry in
InitialNonFlCompVariantData should be used to re-initiate the entry.

Change-Id: Ic36ab70237cb892f3a79d2a3211da2312adedd37
BCP: 2170178047
CR-Id: 002075125800002260422021
(cherry picked from commit 6399cb4d384146ffbc3a61707634cfd8c6ac2e1a)