[FIX]sap.m.Input: Input with OSP sample update
The suggest logic in the sample is now executed every time a suggest event
is fired. This way the items are always filtered on the backend and the
internal input filtering is not affecting the results.
The example will always show only the items returned as a result of the
backend filtration.

Change-Id: I70bcaf8f9e38bd21e69210b0ff910ea26ea101b9
BCP: 002075129500005182052021