[INTERNAL] Core: Disable CustomizingConfiguration
Since 1.95, the CustomizingConfiguration is functionally irrelevant.
No framework coding relies on this module anymore.

With this change the CustomizingConfiguration module is fully disabled,
but for now not deleted.
The empty module is still kept for compatibility reasons.
An error message is logged for all applications which still depend
on this private module.

Change-Id: Id724c3bab0c4fe860d96a3922da02266363ed3cd
JIRA: CPOUI5FRAMEWORK-297