[FIX] sap.f.IllustrationPool: Removed redundant file request
Problem: We require non-existing file which results in console error
and redundant request.

Root Cause: We always require Patterns file from any of the illustration
sets. Since some of the sets don't have their own Patterns file (they
use the default one), we end up with requiring non-existing file.

Solution: We introduce new property in the metadata.json.
The property is called "requireCustomPatterns". If the property is set to
true, then the IllustrationPool will try to load the Patterns file
for the corresponding illustration set.

Note:
The patterns file should be named the following in order to be registered
by the IllustrationPool:
<name of the illustration set>-Patterns

BCP: 002075129500005735492021
Change-Id: I4844ffbdbf3323c8c5f840b7867d08b2b892248f